Make BitStreamReader internal.
Make PatternScanner.Find return an Int32.

+++ Trinity.Encore.Game/IO/PatternScanner.cs
@@ -21,7 +21,7 @@ namespace Trinity.Encore.Game.IO
             _data = data;
         }
 
-        public long? Find(byte?[] pattern)
+        public int? Find(byte?[] pattern)
         {
             Contract.Requires(pattern != null);
